Use colors of the official client
The roads now look a lot better as they aren't as bright anymore and the portal has a much nicer blue.

It's probably a good idea to use some brighter red for the keeperLair for visibility reasons and change the current orange of invaders and source keepers to some kind of red to match the keeperLair.