This makes uses of a PRNG to create more diverse colors
The previous version had the issue that similar user ids could easily result in similar colors. This approach is now using `seedrandom` with the user id to create more dissimilar color seeds.
It is by no means a perfect solution and similar colors are still absolutely possible, but this heavily reduces them for current test cases.